pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "us-stock-gap-dashboard"
version = "0.1.0"
description = "TimesFM + Alpaca: overnight gap risk dashboard (close→next open, 5 trading days)."
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "streamlit>=1.28",
    "python-dotenv>=1.0",
    "alpaca-py>=0.21.0",
    "pytz>=2024.1",
    "torch>=2.1",
    "timesfm>=1.2.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "lightgbm>=4.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = ["ruff>=0.4.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
